+++ src/decorators/index.ts
@@ -1,5 +1,5 @@
 export { VueController } from './controller'
 export { Module } from './module'
 export { Prop } from './prop'
-export { Get, Del } from './request'
+export { Get, Del, Head, Post, Put } from './request/index'
 export { Watch, WatchOption } from './watch'