+++ src/decorators/request/_prefix.ts
@@ -1,0 +1,9 @@
+export function getPrefix(key, options): string {
+  if (typeof options === 'string') {
+    return options;
+  }
+  if (typeof options === 'object' && !!options['path']) {
+    return options['path']
+  }
+  return key;
+}