+++ src/decorators/request/post.ts
@@ -1,4 +1,5 @@
-import { DeveloperUtils } from '../decorator.conf'
+import * as classifier from "../../constants/request_classifier";
+import { getPrefix } from "./_prefix";
 
 export function Post(target: any)
 export function Post(first: any, second?: string) {
@@ -9,13 +10,11 @@ export function Post(first: any, second?: string) {
 }
 
 function propDecorator(options?: any) {
-  return function(target: any, key: string) {
-    DeveloperUtils.decoratorStart();
-    if (!target.$$methods) target.$$methods = {};
 
-    // console.log('Target::', target, key, options);
+  return function (target: any, key: string) {
+    if (!target.$$methods) target.$$methods = {};
     
-    target.$$methods[key] = target[key];
-    DeveloperUtils.decoratorStop();
+    target.$$methods[`${classifier.TYPES.POST}${classifier.SEPARATOR}${getPrefix(key, options)}`] = target[key];
+    return target;
   }
 }