+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Servue } from './servue'
 export * from './decorators'
 export { Server } from './server'
-export { Del, Get } from './decorators/request/index'
+export { Del, Get, Head, Post, Put } from './decorators/request/index'
 export { DeveloperUtils } from './decorators/decorator.conf'